Moved JWT veryfying
Moved the verification of the JWT to the API gateway.
