+++ Implementation/MicroserviceBackend/APIGateway/package.json
@@ -15,12 +15,14 @@
   "dependencies": {
     "@types/jsonwebtoken": "^9.0.9",
     "axios": "^1.9.0",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.23.0",
+    "@types/cookie-parser": "^1.4.8",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.1",
     "@types/jest": "^29.5.14",
