+++ Implementation/MicroserviceBackend/APIGateway/src/app.ts
@@ -5,19 +5,18 @@ import cors from 'cors';
 import authRoutes from './routes/authRoutes';
 import userRoutes from './routes/userRoutes';
 import axios from 'axios';
+import cookieParser from 'cookie-parser';
 
 dotenv.config();
 
 const app = express();
-
+app.use(cookieParser());
 app.use(cors({
     origin: 'http://localhost:5173', // or use '*' to allow all (less secure)
     credentials: true, // allow cookies if needed
 }));
 
 app.use(express.json());
-
-
 
 app.use(authRoutes);
 app.use(userRoutes);
