+++ Implementation/MicroserviceBackend/AuthenticationService/src/tests/jwtMiddleware.test.ts
@@ -1,7 +1,7 @@
-import { verifyJwt } from "../middleware/jwtMiddleware";
+import { verifyJwt } from "../../../APIGateway/src/middleware/jwtMiddleware";
 import { Request, Response, NextFunction } from "express";
 import  jwt, { VerifyErrors }  from "jsonwebtoken";
-import { refreshToken } from "../middleware/refreshToken";
+import { refreshToken } from "../../../APIGateway/src/middleware/refreshToken";
 
 jest.mock('jsonwebtoken', () => ({
     verify: jest.fn(),
